chore: Move logic from `guards::Notifier` to `services::Notifier`
