+++ service/src/services/mod.rs
@@ -7,6 +7,7 @@ pub mod auth_service;
 pub mod invitation_service;
 pub mod jwt_service;
 pub(crate) mod live_xmpp_service;
+pub mod notifier;
 pub mod server_ctl;
 pub mod server_manager;
 pub mod user_service;
